Get rid of SignInDeserializer, it's overkill.
