+++ test/org/sagebionetworks/bridge/models/accounts/SignInTest.java
@@ -7,6 +7,8 @@ import org.junit.Test;
 import org.sagebionetworks.bridge.TestUtils;
 import org.sagebionetworks.bridge.json.BridgeObjectMapper;
 
+import com.fasterxml.jackson.databind.JsonNode;
+
 public class SignInTest {
 
     @Test
@@ -50,4 +52,41 @@ public class SignInTest {
         assertEquals("email@email.com", signIn.getEmail());
         assertEquals("myReauthToken", signIn.getReauthToken());
     }
+    
+    @Test
+    public void test() throws Exception {
+        JsonNode node = BridgeObjectMapper.get().readTree(TestUtils.createJson("{"+
+                "'email':'emailValue',"+
+                "'password':'passwordValue',"+
+                "'study':'studyValue',"+
+                "'token':'tokenValue',"+
+                "'phone':'1234567890',"+
+                "'reauthToken':'reauthTokenValue'"+
+                "}"));
+        
+        SignIn signIn = BridgeObjectMapper.get().readValue(node.toString(), SignIn.class);
+        assertEquals("emailValue", signIn.getEmail());
+        assertEquals("passwordValue", signIn.getPassword());
+        assertEquals("studyValue", signIn.getStudyId());
+        assertEquals("tokenValue", signIn.getToken());
+        assertEquals("1234567890", signIn.getPhone());
+        assertEquals("reauthTokenValue", signIn.getReauthToken());
+    }
+    
+    @Test
+    public void acceptsUsernameAsEmail() throws Exception {
+        JsonNode node = BridgeObjectMapper.get().readTree(TestUtils.createJson("{"+
+                "'username':'emailValue',"+
+                "'password':'passwordValue',"+
+                "'study':'studyValue',"+
+                "'token':'tokenValue',"+
+                "'reauthToken':'reauthTokenValue'"+
+                "}"));
+        SignIn signIn = BridgeObjectMapper.get().readValue(node.toString(), SignIn.class);
+        assertEquals("emailValue", signIn.getEmail());
+        assertEquals("passwordValue", signIn.getPassword());
+        assertEquals("studyValue", signIn.getStudyId());
+        assertEquals("tokenValue", signIn.getToken());
+        assertEquals("reauthTokenValue", signIn.getReauthToken());
+    }    
 }
